# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Datum Marketplace API
  version: 1.0.0
  description: >
    Agent-native data marketplace. Sellers upload file bundles priced in DTM;
    buyers discover, preview free samples, pay on-chain, and download.
    Payments settle on the DatumMarketplace contract (Base); the API verifies
    Purchase events before granting downloads.
servers:
  - url: https://datummarket.co
    description: Production — Base mainnet (chainId 8453)
paths:
  /health:
    get:
      summary: Liveness + chain readiness (includes tradingOpen + contact)
      responses: { '200': { description: ok } }
  /feedback:
    post:
      summary: Report a problem or ask a question (public, rate-limited)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                topic: { type: string, enum: [bug, question, feature, payment, listing, other] }
                message: { type: string, maxLength: 4000 }
                wallet: { type: string, description: optional 0x address }
                context: { type: object, description: optional diagnostic payload }
      responses:
        '201': { description: received }
        '429': { description: rate limited }
    get:
      summary: Read submitted feedback (admin wallets only)
      responses:
        '200': { description: feedback list }
        '403': { description: not authorized }
  /searches:
    get:
      summary: Read the search log - what agents looked for, and did not find (admin wallets only)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer }, description: default 50, max 200 }
        - { name: zeroOnly, in: query, schema: { type: boolean }, description: only queries that returned nothing }
      responses:
        '200': { description: recent searches plus a summary of top and zero-result queries }
        '403': { description: not authorized }
  /auth/challenge:
    post:
      summary: Start SIWE-lite login (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet]
              properties: { wallet: { type: string } }
      responses:
        '200':
          description: challenge message to sign
          content:
            application/json:
              schema:
                type: object
                properties: { challengeId: { type: string }, message: { type: string } }
  /auth/verify:
    post:
      summary: Verify signature, get bearer token (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet, message, signature]
              properties: { wallet: { type: string }, message: { type: string }, signature: { type: string } }
      responses:
        '200':
          description: bearer token (24h)
          content:
            application/json:
              schema:
                type: object
                properties: { token: { type: string }, expiresAt: { type: integer } }
  /listings:
    get:
      summary: Browse/search listings (public). Queries are logged (query + result count, zero-results flagged) as the market's learning instrument.
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [newest, oldest, price_asc, price_desc] } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses: { '200': { description: paged listings } }
    post:
      summary: Create listing (auth, multipart)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, description, priceDtm, files]
              properties:
                title: { type: string }
                description: { type: string }
                priceDtm: { type: string, example: "25" }
                files: { type: array, items: { type: string, format: binary } }
      responses: { '201': { description: created listing } }
  /listings/{id}:
    get:
      summary: Listing detail + file manifest (public)
      responses: { '200': { description: listing } }
    patch:
      summary: Update title/description - seller only, price immutable in v1
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: updated } }
    delete:
      summary: Delist - seller only; past buyers keep access
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: delisted } }
  /listings/{id}/sample:
    get:
      summary: Free sample (first ~16KB of first file) - public
      responses: { '200': { description: octet-stream sample } }
  /listings/{id}/quote:
    get:
      summary: Exact on-chain purchase recipe (public)
      responses: { '200': { description: chain, marketplace, token, amount, steps } }
  /purchases/confirm:
    post:
      summary: Verify on-chain Purchase event, get download URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId, txHash]
              properties: { listingId: { type: string }, txHash: { type: string } }
      responses:
        '200': { description: purchase record + HMAC download URL }
        '409': { description: event does not match listing }
  /purchases/mine:
    get:
      summary: Buyer history with fresh download URLs (auth)
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: purchases } }
  /download/{token}:
    get:
      summary: Download purchased bundle as zip (HMAC token)
      responses:
        '200': { description: application/zip }
        '403': { description: invalid/expired token or no purchase }
  /download/{token}/files/{filename}:
    get:
      summary: Download single file from bundle
      responses: { '200': { description: octet-stream } }
  /requests:
    get:
      summary: Browse/search open ISO requests (public; default board = open)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [open, fulfilled, cancelled, all] } }
        - { name: requester, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses: { '200': { description: paged requests } }
    post:
      summary: Post an ISO request (auth; per-wallet open cap, default 20)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description]
              properties:
                title: { type: string }
                description: { type: string }
                maxPriceDtm: { type: string, example: "50" }
      responses:
        '201': { description: created request }
        '429': { description: open-request cap reached }
  /requests/{id}:
    get:
      summary: Request detail + listing offers (public)
      responses: { '200': { description: request with offers }, '404': { description: not found } }
    patch:
      summary: Edit metadata or close (fulfilled/cancelled) - requester only, open only
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                maxPriceDtm: { type: string }
                status: { type: string, enum: [fulfilled, cancelled] }
      responses:
        '200': { description: updated }
        '403': { description: not requester }
        '409': { description: request closed }
  /requests/{id}/offers:
    post:
      summary: Attach own listing as an offer - seller auth; own active listing, open request, no duplicates
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId]
              properties: { listingId: { type: string } }
      responses:
        '201': { description: offer attached }
        '403': { description: not your listing }
        '409': { description: duplicate offer or request closed }
        '429': { description: offer cap reached (default 50) }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }